Fix erranous renaming of util function file
