+++ client/pages/clubs/[id].tsx
@@ -3,7 +3,7 @@ import type { Theme } from '@mui/material';
 import { GetServerSidePropsContext, InferGetServerSidePropsType } from 'next';
 import { useRouter } from 'next/router';
 import { getClub } from '../../src/api';
-import { getAccessLevel, getParams } from '../../src/util/miscUtilUtil';
+import { getAccessLevel, getParams } from '../../src/util/miscUtil';
 import { darkSwitchGrey } from '../../src/util/cssUtil';
 
 import Container from '@mui/material/Container';
