+++ client/src/components/shared/paragraph.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { parseLinks } from '../../util/miscUtilUtil';
+import { parseLinks } from '../../util/miscUtil';
 
 import Box from '@mui/material/Box';
 import Typography from '@mui/material/Typography';
